requirements.txt

Django==3.0.4
djangorestframework==3.11.0

